Changes of version: 20111223

+++ libuna/libuna_base64_stream.h
@@ -39,7 +39,7 @@ LIBUNA_INLINE \
 int libuna_base64_character_copy_to_sixtet(
      uint8_t base64_character,
      uint8_t *base64_sixtet,
-     uint8_t base64_variant,
+     uint32_t base64_variant,
      liberror_error_t **error );
 
 LIBUNA_EXTERN \
@@ -50,7 +50,7 @@ int libuna_base64_triplet_copy_from_base64_stream(
      size_t base64_stream_size,
      size_t *base64_stream_index,
      uint8_t *padding_size,
-     uint8_t base64_variant,
+     uint32_t base64_variant,
      liberror_error_t **error );
 
 LIBUNA_EXTERN \
@@ -61,7 +61,7 @@ int libuna_base64_triplet_copy_to_base64_stream(
      size_t base64_stream_size,
      size_t *base64_stream_index,
      uint8_t padding_size,
-     uint8_t base64_variant,
+     uint32_t base64_variant,
      liberror_error_t **error );
 
 LIBUNA_EXTERN \
@@ -89,7 +89,7 @@ int libuna_base64_stream_size_to_byte_stream(
      uint8_t *base64_stream,
      size_t base64_stream_size,
      size_t *byte_stream_size,
-     uint8_t base64_variant,
+     uint32_t base64_variant,
      uint8_t flags,
      liberror_error_t **error );
 
@@ -99,7 +99,7 @@ int libuna_base64_stream_copy_to_byte_stream(
      size_t base64_stream_size,
      uint8_t *byte_stream,
      size_t byte_stream_size,
-     uint8_t base64_variant,
+     uint32_t base64_variant,
      uint8_t flags,
      liberror_error_t **error );
 
@@ -108,7 +108,7 @@ int libuna_base64_stream_size_from_byte_stream(
      uint8_t *byte_stream,
      size_t byte_stream_size,
      size_t *base64_stream_size,
-     uint8_t base64_variant,
+     uint32_t base64_variant,
      liberror_error_t **error );
 
 LIBUNA_EXTERN \
@@ -117,7 +117,7 @@ int libuna_base64_stream_copy_from_byte_stream(
      size_t base64_stream_size,
      uint8_t *byte_stream,
      size_t byte_stream_size,
-     uint8_t base64_variant,
+     uint32_t base64_variant,
      liberror_error_t **error );
 
 #if defined( __cplusplus )
